Level select is now accessible from it's own button. Game Options panel mockup done. Code documented thoroughly.
