+++ pom.xml
@@ -4,7 +4,7 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
 
-    <groupId>org.javacodephoenixsociety</groupId>
+    <groupId>org.jcps</groupId>
     <artifactId>Vehicle2-redux</artifactId>
     <version>1.0-SNAPSHOT</version>
     <pluginRepositories>
@@ -34,7 +34,7 @@
                 <configuration>
                     <archive>
                         <manifest>
-                            <mainClass>Vehicle2</mainClass>
+                            <mainClass>org.jcps.vehicle2redux.V2RMain</mainClass>
                         </manifest>
                     </archive>
                 </configuration>
