--- src/main/java/TriggerEvent.java
@@ -1,14 +1,0 @@
-import java.util.EventObject;
-
-public class TriggerEvent extends EventObject {
-    private final String message;
-
-    public TriggerEvent(Object source, String message) {
-        super(source);
-        this.message = message;
-    }
-
-    public String getMessage() {
-        return message;
-    }
-}
