--- src/main/java/TriggerListener.java
@@ -1,5 +1,0 @@
-import java.util.EventListener;
-
-public interface TriggerListener extends EventListener {
-    void onEventOccurred(TriggerEvent event);
-}
