+++ src/main/java/org/jcps/vehicle2redux/V2RMain.java
@@ -1,5 +1,8 @@
+package org.jcps.vehicle2redux;
+
 import javax.swing.*;
 import java.awt.*;
+import java.awt.event.ActionListener;
 import java.awt.event.KeyEvent;
 import java.awt.event.WindowEvent;
 import java.awt.event.WindowListener;
@@ -10,25 +13,79 @@ import java.awt.event.WindowListener;
  */
 public class V2RMain extends JFrame implements TriggerListener {
     // Constants for the game state
+    /**
+     * Represents the game state when at the main menu panel.
+     */
     public final static int MENU_PANEL = 0;
+
+    /**
+     * Represents the game state when at the game panel.
+     */
     public final static int GAME_PANEL = 1;
+
+    /**
+     * Represents the game state when at the options panel.
+     */
     public final static int OPTIONS_PANEL = 2;
 
+    /**
+     * Represents the game state when at the level selection panel.
+     */
+    public final static int SELECT_PANEL = 3;
+
     // General variables
+
+    /**
+     * The current game state.
+     */
     public static int gameState;
-    public Vehicle2 objV2;
+
+    /**
+     * The main Vehicle2 instance for the game.
+     */
+    public Vehicle2 vehicle2r;
+
+    /**
+     * The previous level in the game.
+     */
     private int prevLevel;
+
+    /**
+     * The Timer instance for handling game events.
+     */
     private static Timer timer = null;
-    private static MenuPanel mp;
-    private static LevelSelectPanel lsp;
+
+    /**
+     * The MenuPanel instance for the main menu.
+     */
+    private final MenuPanel mp;
+
+    /**
+     * The LevelSelectPanel instance for level selection.
+     */
+    private LevelSelectPanel levelSelectPanel;
+
+    /**
+     * The OptionsPanel instance for managing the game options.
+     */
+    private final OptionsPanel optionsPanel;
+
+    /**
+     * The delay interval for the timer (in milliseconds).
+     */
     private final int delay = 10;
+
+    /**
+     * Static debug flag. Set to true to show debugging information.
+     */
+    public static boolean DEBUG = true;
 
     /**
      * Constructs the V2RMain frame for the game.
      */
     public V2RMain() {
         setTitle("V2: Redux");
-        setUndecorated(false);
+        setUndecorated(true);
         setSize(800, 600);
         setDefaultCloseOperation(JFrame.DO_NOTHING_ON_CLOSE);
         setLocation(0, 0);
@@ -40,7 +97,9 @@ public class V2RMain extends JFrame implements TriggerListener {
             update(); // Method to update game state
         });
         timer.setInitialDelay(20);
-        mp = new MenuPanel();
+
+        this.mp = new MenuPanel();
+        this.optionsPanel = new OptionsPanel();
 
         // Add window listener to handle closing action
         addWindowListener(new WindowListener() {
@@ -95,8 +154,8 @@ public class V2RMain extends JFrame implements TriggerListener {
     public static void main(String[] args) {
         Dimension screenSize = Toolkit.getDefaultToolkit().getScreenSize();
         V2RMain v2r = new V2RMain();
-        v2r.objV2 = new Vehicle2();
-        lsp = new LevelSelectPanel(v2r.objV2.maps);
+        v2r.vehicle2r = new Vehicle2();
+        v2r.levelSelectPanel = new LevelSelectPanel(v2r.vehicle2r.maps);
 
         // Calculate the center position
         int centerX = (screenSize.width - 800) / 2;
@@ -105,33 +164,36 @@ public class V2RMain extends JFrame implements TriggerListener {
         // Set the frame location
         v2r.setLocation(centerX, centerY);
 
-        v2r.objV2.addEventListener(v2r);
-        v2r.lsp.addEventListener(v2r);
-        v2r.add(mp, BorderLayout.CENTER);
+        v2r.vehicle2r.addEventListener(v2r);
+        v2r.levelSelectPanel.addEventListener(v2r);
+        v2r.optionsPanel.addEventListener(v2r);
+        v2r.add(v2r.mp, BorderLayout.CENTER);
 
         gameState = MENU_PANEL;
         v2r.revalidate();
         v2r.repaint();
     }
 
-    /** Method to update game logic */
+    /**
+     * Method to update game logic
+     */
     private void update() {
         if (gameState == GAME_PANEL) {
-            if (objV2.runState) {
-                objV2.run();
+            if (vehicle2r.runState) {
+                vehicle2r.run();
             } else {
                 // Stop the timer from running a duplicate loop
                 timer.stop();
 
                 if (prevLevel != -2) {
-                    if (prevLevel < objV2.currentLevel) {
-                        ++objV2.currentLevel;
+                    if (prevLevel < vehicle2r.currentLevel) {
+                        ++vehicle2r.currentLevel;
                         prevLevel++;
                     }
                 }
-                if (objV2.currentLevel < objV2.maps.size()) {
-                    if (objV2.gp != objV2.maps.get(objV2.currentLevel).lp) {
-                        objV2.init();
+                if (vehicle2r.currentLevel < vehicle2r.maps.size()) {
+                    if (vehicle2r.gp != vehicle2r.maps.get(vehicle2r.currentLevel).lp) {
+                        vehicle2r.init();
                         timer.start();
                     } else {
                         returnToMenu();
@@ -149,12 +211,13 @@ public class V2RMain extends JFrame implements TriggerListener {
         repaint();
     }
 
-    /** Returns to the main menu.
-    */
+    /**
+     * Returns to the main menu.
+     */
     private void returnToMenu() {
         gameState = MENU_PANEL;
         getContentPane().removeAll();
-        add(mp);
+        add(this.mp);
     }
 
     /**
@@ -162,11 +225,11 @@ public class V2RMain extends JFrame implements TriggerListener {
      */
     public void startGame() {
         // Simple protection for no maps
-        if (objV2.maps.size() != 0) {
+        if (vehicle2r.maps.size() != 0) {
             this.getContentPane().removeAll();
-            this.add(objV2);
+            this.add(vehicle2r);
             gameState = GAME_PANEL;
-            objV2.init();
+            vehicle2r.init();
             timer.start();
         } else {
             JOptionPane.showMessageDialog(null,
@@ -176,33 +239,39 @@ public class V2RMain extends JFrame implements TriggerListener {
         }
     }
 
-    /** Processes events fired from children.
+    /**
+     * Processes events fired from children.
      *
      * @param event Incoming event from trigger
      */
     @Override
     public void onEventOccurred(TriggerEvent event) {
-        System.out.println("Event occurred: " + event.getMessage());
+        if (DEBUG) System.out.println("Event occurred: " + event.getMessage());
         if (event.getMessage().equals("exit_loop")) {
             if (gameState == MENU_PANEL) {
                 System.exit(0);
             }
-
+            if (gameState == OPTIONS_PANEL || gameState == SELECT_PANEL) {
+                timer.stop();
+                gameState = MENU_PANEL;
+                this.getContentPane().removeAll();
+                this.add(this.mp);
+            }
             if (gameState == GAME_PANEL) {
                 timer.stop();
                 gameState = MENU_PANEL;
-                objV2.currentLevel = 0;
+                vehicle2r.currentLevel = 0;
                 prevLevel = -1;
                 this.getContentPane().removeAll();
-                this.add(mp);
+                this.add(this.mp);
             }
         }
         if (event.getMessage().contains("map:")) {
             String msg = event.getMessage();
             msg = msg.substring(msg.lastIndexOf(':') + 1);
 
-            System.out.println("Go to map: " + msg);
-            objV2.currentLevel = Integer.parseInt(msg);
+            if (DEBUG) System.out.println("Go to map: " + msg);
+            vehicle2r.currentLevel = Integer.parseInt(msg);
             prevLevel = -2;
             startGame();
         }
@@ -212,14 +281,22 @@ public class V2RMain extends JFrame implements TriggerListener {
     }
 
     /**
-     * Opens the options menu (Temporary the level select menu).
+     * Opens the options menu.
      */
     private void openOptions() {
+        this.getContentPane().removeAll();
+        this.add(this.optionsPanel);
+        gameState = OPTIONS_PANEL;
+        revalidate();
+        repaint();
+    }
+
+    private void openLevelSelect() {
         // Simple protection for no maps
-        if (objV2.maps.size() != 0) {
+        if (vehicle2r.maps.size() != 0) {
             this.getContentPane().removeAll();
-            this.add(lsp);
-            gameState = OPTIONS_PANEL;
+            this.add(this.levelSelectPanel);
+            gameState = SELECT_PANEL;
         } else {
             JOptionPane.showMessageDialog(null,
                     "Can not start the game. There are no map files in the Levels/ directory",
@@ -234,46 +311,75 @@ public class V2RMain extends JFrame implements TriggerListener {
      * Represents a panel containing menu buttons for starting, configuring options, and exiting the game.
      */
     public class MenuPanel extends JPanel {
+        /**
+         * Button to initiate the start of the game.
+         */
         JButton btnStart;
+
+        /**
+         * Button to exit the game.
+         */
         JButton btnExit;
+
+        /**
+         * Button to access and configure game options.
+         */
         JButton btnOptions;
+
+        /**
+         * Button to navigate to the level selection panel.
+         */
+        JButton btnLevelSelect;
 
         /**
          * Constructs a MenuPanel with buttons for starting the game, accessing options, and exiting the game.
          */
         public MenuPanel() {
-            // Set up GridBagLayout for the frame
-            setLayout(new GridBagLayout());
+            Dimension buttonSize = new Dimension(200, 40);
             GridBagConstraints gbc = new GridBagConstraints();
             gbc.gridx = 0;
-            gbc.gridy = 0;
-            gbc.weightx = 1.0;
-            gbc.weighty = 1.0;
-            gbc.fill = GridBagConstraints.CENTER;
+            gbc.weightx = 1.0; // Occupy full width
+            gbc.fill = GridBagConstraints.RELATIVE;
 
-            // Create a nested JPanel with FlowLayout to center the buttons
-            JPanel centeredPanel = new JPanel();
-            centeredPanel.setLayout(new BoxLayout(centeredPanel, BoxLayout.Y_AXIS));
+            setLayout(new GridBagLayout());
 
-            btnStart = new JButton("Start Game");
-            btnStart.setMnemonic(KeyEvent.VK_S);
-            btnStart.addActionListener(e -> startGame());
+            // Set up "Start Game" button
+            btnStart = createButton("Start Game", buttonSize, KeyEvent.VK_S, e -> startGame());
 
-            btnOptions = new JButton("Options (LS TEMP)");
-            btnOptions.setMnemonic(KeyEvent.VK_O);
-            btnOptions.addActionListener(e -> openOptions());
+            // Set up Level Select button
+            btnLevelSelect = createButton("Level Select", buttonSize, KeyEvent.VK_L, e -> openLevelSelect());
 
-            btnExit = new JButton("Exit Game");
-            btnExit.setMnemonic(KeyEvent.VK_X);
-            btnExit.addActionListener(e -> {
+            // Set up Options button
+            btnOptions = createButton("Options", buttonSize, KeyEvent.VK_O, e -> openOptions());
+
+            // Set up exit button
+            btnExit = createButton("Exit Game", buttonSize, KeyEvent.VK_X, e -> {
                 LevelUtilities.writeHashmapToFile(Vehicle2.levelTimes, "Level-times.txt");
                 System.exit(0);
             });
 
-            centeredPanel.add(btnStart);
-            centeredPanel.add(btnOptions);
-            centeredPanel.add(btnExit);
-            add(centeredPanel, gbc);
+            // Add to panel
+            add(btnStart, gbc);
+            add(btnLevelSelect, gbc);
+            add(btnOptions, gbc);
+            add(btnExit, gbc);
+        }
+
+        /**
+         * Creates and configures a JButton with the specified text, size, mnemonic, and ActionListener.
+         *
+         * @param text     The text to be displayed on the button.
+         * @param size     The preferred size of the button.
+         * @param mnemonic The mnemonic key code for keyboard shortcut.
+         * @param action   The ActionListener to be triggered when the button is clicked.
+         * @return A configured JButton with the specified properties.
+         */
+        private JButton createButton(String text, Dimension size, int mnemonic, ActionListener action) {
+            JButton button = new JButton(text);
+            button.setPreferredSize(size);
+            button.setMnemonic(mnemonic);
+            button.addActionListener(action);
+            return button;
         }
     }
 }
