+++ src/main/resources/META-INF/MANIFEST.MF
@@ -1,3 +1,3 @@
 Manifest-Version: 1.0
-Main-Class: Vehicle2
+Main-Class: org.jcps.vehicle2redux.V2RMain
